third_party/blink/web_tests/fast/parser/fragment-parser-expected.txt

Test that fragment parsing does not affect the host document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS parse('<span><body bgcolor=red>') is ['<span></span>','<html><body></body></html>']
PASS parse('<span><html bgcolor=red>') is ['<span></span>','<html><body></body></html>']
PASS parse('<span><meta>') is ['<span><meta></span>','<html><body></body></html>']
PASS parse('<span><base>') is ['<span><base></span>','<html><body></body></html>']
PASS parse('<html><script>') is ['<script></script>','<html><body></body></html>']
PASS parse('<html><style>') is ['<style></style>','<html><body></body></html>']
PASS parse('<html><meta>') is ['<meta>','<html><body></body></html>']
PASS parse('<html><link>') is ['<link>','<html><body></body></html>']
PASS parse('<html><object>') is ['<object></object>','<html><body></body></html>']
PASS parse('<html><embed>') is ['<embed>','<html><body></body></html>']
PASS parse('<html><title>') is ['<title></title>','<html><body></body></html>']
PASS parse('<html><isindex>') is ['<isindex></isindex>','<html><body></body></html>']
FAIL parse('<html><base>') should be ,<html><body></body></html>. Was <base>,<html><body></body></html>.
PASS parse('<html><div>') is ['<div></div>','<html><body></body></html>']
PASS parse('<frameset>') is ['','<html><body></body></html>']
PASS parse('<html>x', true) is ['x','no document element']
PASS successfullyParsed is true

TEST COMPLETE
